Refactor doctests to be runnable by pytest
From now on, doctests will be expected to be run by pytest. This means
that the sphinx directive

.. testsetup:: [group]

is no longer allowed; replace it with

.. doctest:: [group]
   :hide:

or simply

.. doctest::
   :hide:

Each module containing doctests should run with the cwd changed to a
temporary directory, thanks to the new cwd_testing() fixture
(conftest.py).