Implement DefineHandler and AddMissingSectionHandler (task #156363)
We automatically instantiate DefineHandlers for each [define/*] config
section when the Config object is instantiated. I am not terribly fond
of hard-wiring the instantiation of handlers in the Config constructor,
but at least it can be turned off with a flag when it is not needed. For
the moment, this seems to be the simplest solution.