Refactor accessor based on DataResponses and IndexResponses classes
The goal is to separate data themselves from the choice to be made on
indices.
DataResponses contains the usual nested list of responses and a
flattened one. Flat one means score identifiers are sent at the same
level as the other response characteristics. A correspondance dictinary
is built at the same time.
IndexResponses can accept nested or flat list to build the index.
Selections are done on top of that. Only sets or "dsets" objets can be
retrieved.
Accessor has been updated accordingly.

Change name or ResponseData in DataResponses for more consistency with
IndexResponses and bug correction.