Flatten the result in parsing instead of in accessor.
Flattening the structure of the result has been moved to
pyparsing_t4/transform.py in order to get an easier and more general
structure in accesor. As a consequence, the Response namedtuple has also
been removed. Results are now stored in a standard dictionary under the
key 'results' while the type of results is stored under the key 'result_type'.
This last one is meant for debugging and testing.
Score and response indices are added and can be used as selection keys (for
example when no score name is available).
Composition details (temperature, composition, etc) are now transformed in a
dictionary of tuples that can be directly selected from index. This
modification implies a small manipulation at the final construction of the
response dictionary to remove the key 'compos_details' and get directly access
to the content of the dictionary under it (temperature, etc).
Tests, cleaning and pylint/flake8 fixes.