Change structure of tests: move tests to Test class.
Implement new Test classes test specific that store the datasets and the
criterions. They also perform the test via the evaluate method, returning
a specific TestResult. The idea of functions arguments is for the moment
forgotten, but we can go back to it (seemed easier to implement in this
way, but...) TestResults classes have now a default argument that is the
test himself, it only returns its result.