Implement comments.
Doc: links to list in complex return types.
Remove some of the exceptions to LOGGER message and empty return
(dimension calculation in plot_elements).
Remove the special cas of one curve and adpat the multiple one to
deal with it.
Remove static methods returning the matplotlib.pyplot instance (can
be directly obtained with import by the user) and add the exception
in XRepresentation call.