Use an index to determine the curve style in matplotlib plots.
Curves with the same index will have the same style (color, markers).
When the PlotTemplate comes from a TestResult this index corresponds to
the index of the tested dataset in the list of datasets. Reference is 0.
In case of use of the join function or method, if a plot is given with a
new y-axis (so different yname), it is considered the user put the correct
index. If 2 PlotTemplates are joined with the same ynames, the index are
incremented by the number of curves in the first PlotTemplate.