Update of the documentation and fix the pdf output.
Among the updates of the documentation:
 - splitting in rst files of the documentation specific to Tripoli-4
   parsing in eponine allowing tocs and more suitable levels of titles;
 - updates of documentation of spectra and meshes that was not consistent
   with the code anymore.
 - use field lists instead of items in doc
Fix the pdf output: use xelatex per default to take into account the
unicode characters used in documentation (Greek characters, cute little
stars, ...). For the stars, not all the fonts take them into account
even if the compilation is not failing (DejaVu is fine).