Documentation without tests and only for developers.
Documentation can only be built by developers (pip installation with
the '-e' option and [dev]). The tests, run with pytest, have been
removed from the documentation coverage even if the docstrings still
exist and have to be written. It is always possible to build a small
parallel documentation for them. The only quotation of tests (in
developers part of the documentation) has been changed to not point
on it (as no link possible).