Adapt default keff estimation to ResponsesBook.
Default keff estimation or "best keff estimation", now named
auto_keff in the valjean (referring to automatic, or Tripoli-4,
calculation of the number of discarded batches to get the "best"
estimation of keff).
Datasets can now be obtained thanks to data_convertor.
used_batch_res also in auto keff and kij, data_convertor for it