Fix test_run integration test
This test has been broken since... forever? The integration tests were
generating a bunch of test jobs with the same name. Since the name of
the test module was the same for all the jobs, Python was only importing
the first one and assuming all the rest of the modules were identical,
which was actually skipping a lot of tests!