Create all output paths before scheduling tasks
This commit fixes a race condition in the creation of the '*-root'
directories at the beginning of the run. For instance, if the
'checkout-root' directory does not exist and several checkout tasks try
to create it at the same time, they race for creating it and some of
them might fail.

With this commit, we also ensure that the path to the log file exists
when creating it.