Change logic to know if job run in parallel and management of times.
Now a parallel job is identified by the presence of the string
"number of tasks is" in the beginning of TRIPOLI-4 output.
All times are stored instead of only the last one, so simulation time
can be accessed for each batch even if parsing of all batches is not
required. This was mainly useful for the elapsed time, in parallel job,
that is not parsed (as the output string is normally cut at simulation
time).
A check is performed on the additional end_flag (times) to only store
the ones linked to edited batch results (else simulation time now appears
for the 20 first batches...).
CAUTION: one failing test in test_scan_t4 (test_tt_simple_packet20_para)
due to new selection on para and times in scan. Corrected in next commit.