Add scan variables in ResponseBook.globals
Add a condition on end flags (usually times) as the batch number and the
simulation time now seem to allways appear in the TRIPOLI-4 output run in
MONO-processor mode for the 20 first batches, meaning the number of times
'simulation time' appaers does not correspond to the number of edited
batches.
Check on the times added in the test_parsing for more general check.
Fixes in use and accessor as this last one now takes as argument a T4Parser
instead of the T4Parser.result (to get access to the global variables).