Change logic of ResponseBook length calculation.
The length of a ResponseBook is per default the number of responses
it contains, but if no responses where available it returns the length
of the globals. This change is done mainly to deal in a different way
with the if ResponseBook, that returns the length of ResponseBook. In
some cases we could only have globals (of course ResponseBook might
not be useful in that case).