Add a TestResultFailed with its relative rST.
Use a RstTextTemplate that is directly called in representation
(can be changed if not OK...). In that case the plot representer
returns None as the message have normally already be given by the
table representer.
New exception called in tests.
Some corrections done on the tests relative to nature of floats
(32, 64, etc) due to future changes in hypothesis, current warnings,
will be errors.