Major simplification of the CLI interface
This commit gets rid of many of the top-level commands, such as
`checkout`, `build`, `eval`, `report`. All the work is delegated to the
`run` command now, and the fine control of which tasks must be run
(filtering) is delegated to the job script. For this reason, this commit
also removes the PRIORITY mechanism for task filtering.

Also, the `-a` option (arguments for the `job` function) is replaced by
positional arguments. Now instead of running

  $ valjean -j job.py -a arg1 -a arg2 -a arg3 report

you simply do

  $ valjean run job.py arg1 arg2 arg3