Implement a way to pass kwargs to the job() function
This adds a `-k` CLI option to the `valjean run` and `valjean graph`
commands that can be used to pass kwarg-style options to job(). The use
case for this is defining kwarg-only keywords in job() with reasonable
defaults (e.g. system paths), without suppressing the possibility to
modify these parameters.

The `-k` option is tested in a dedicated integration test.