Fix error involving serialization of non-contiguous Numpy arrays
Error was: TypeError: memoryview: casts are restricted to C-contiguous views.
Use it to correct bug: columns can be lists but their fingerprint could not
be determined as lists have no data member. np.require solves that as it
converts lists in arrays.
Add a "test" to be sure fingerprints are also done for lists.