Add possibility to read and use Tripoli-4 depletion results
Minimal version of ROOT: 6.22.00.
This includes the import of the Tripoli-4 depletion post-treatment, given
with Tripoli-4 code. It is based on #cd37c0db, after update to ROOT 6. Some
modifications have been applied to these files:
- correction of Nauchi orthograph
- reindentation
- add new helper methods to retrieve the list of isotopes or the list of reaction
names
- add a new C++ structure, Histogram, to get results as function of burnup or time
independently of TGraphErrors
- add thermal and fast reaction rates separately
The compilation of these C++ classes as a ROOT library is done in a temporary folder
given by the user.
Python equivalent of the C++ methods are automatically generated thanks
to accessors using kwargs (doc and code).
Documentation and test of the module.
Jenkins test only for python3.9.