Refactor TestResultStatsTestsByLabels to avoid uninitialised attributes
The old version of TestResultStatsTestsByLabels had two attributes that
were effectively left uninitialized at construction : self.index and
self.labels_lod. They were both initialized to None, which is dangerous
because it means that the object is left in a half-initialised state.
This commit removes the problematic attributes and computes them when
the test is actually evaluated.