Write out report plots in parallel
This commit uses the multiprocessing module to spawn writers for the
report plots. The number of subprocesses equals the number of worker
threads by default. Combined with the parallel execution of tasks, this
may lead to large system loads. However, the most typical use case for
the report task is to run alone at the end of the job.