Refactor logger
The logger logic has been changed: valjean now has a logger by module.
This allows the user to change the logger level of each module.

A new logger level has been introduced: NOTE, between INFO and DEBUG.
The logger level of some messages changed, in order to get a lower verbosity
in INFO level.
The formatting of all levels are now identical.
Tests are done on the logger, to validate the level change by modules or
sub-packages.