Update packages
Add version requirements in Jenkinsfile (tox dependencies), else failing on
Ubuntu20. The installation from pyproject.toml is working with pip and with
Poetry (but not in python3.6 as it cannot use that version of poetry.lock,
mainly tested with python3.10).
Updates of rstcheck to take into account the upgrade.
Use a different cache directory for each version of python to avoid conflicts
in packages versions (only way found to still be able to test
python 3.6 -> 3.10 using tox).

First adaptation to sphinx update

pyproject working for poetry (but not in python3.6 as cannot use that version of poetry.lock)